// apps/logistics-service/src/module/interfaces/dtos/requests/zone.request.dto.ts
import { z } from 'zod';

export const CreateZoneRequestSchema = z.object({
  code: z.string().min(2).max(20),
  name: z.string().min(2).max(100),
  type: z.string().min(1),
  divisions: z.array(z.string()).default([]),
  districts: z.array(z.string()).default([]),
});

export type CreateZoneRequestDTO = z.infer<typeof CreateZoneRequestSchema>;